p4i integration Integrate 258475 from 4.3 to main: Fixes:    Text showing decimal places regression Task:     155395        RevBy:    janarve Details:  Broke in change #70
[git-p4: depot-paths = "//depot/qt/main/": change = 258476]

+++ src/gui/widgets/qprogressbar.cpp
@@ -428,7 +428,7 @@ QString QProgressBar::text() const
         return result;
     }
 
-    qreal progress = ((qreal(d->value) - qreal(d->minimum)) * 100.0) / totalSteps;
+    int progress = ((qreal(d->value) - qreal(d->minimum)) * 100.0) / totalSteps;
     result.replace(QLatin1String("%p"), QString::fromLatin1("%1").arg(progress));
     return result;
 }
